requirements.txt
streamlit
google-cloud-documentai
pandas
xlsxwriter
openpyxl